Move duplicate code of models.py and models2.py into model_base.py
