+++ modules/models.py
@@ -21,91 +21,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-
-class StdConv2d(nn.Conv2d):
-
-    def forward(self, x):
-        w = self.weight
-        v, m = torch.var_mean(w, dim=[1, 2, 3], keepdim=True, unbiased=False)
-        w = (w - m) / torch.sqrt(v + 1e-10)
-        return F.conv2d(
-            x, w, self.bias, self.stride, self.padding, self.dilation, self.groups
-        )
-
-
-def conv3x3(cin, cout, stride=1, groups=1, bias=False):
-    return StdConv2d(
-        cin, cout, kernel_size=3, stride=stride, padding=1, bias=bias, groups=groups
-    )
-
-
-def conv1x1(cin, cout, stride=1, bias=False):
-    return StdConv2d(cin, cout, kernel_size=1, stride=stride, padding=0, bias=bias)
-
-
-def tf2th(conv_weights):
-    """Possibly convert HWIO to OIHW."""
-    if conv_weights.ndim == 4:
-        conv_weights = conv_weights.transpose([3, 2, 0, 1])
-    return torch.from_numpy(conv_weights)
-
-
-class PreActBottleneck(nn.Module):
-    """Pre-activation (v2) bottleneck block.
-
-    Follows the implementation of "Identity Mappings in Deep Residual Networks":
-    https://github.com/KaimingHe/resnet-1k-layers/blob/master/resnet-pre-act.lua
-
-    Except it puts the stride on 3x3 conv when available.
-    """
-
-    def __init__(self, cin, cout=None, cmid=None, stride=1):
-        super().__init__()
-        cout = cout or cin
-        cmid = cmid or cout // 4
-
-        self.gn1 = nn.GroupNorm(32, cin)
-        self.conv1 = conv1x1(cin, cmid)
-        self.gn2 = nn.GroupNorm(32, cmid)
-        self.conv2 = conv3x3(cmid, cmid, stride)  # Original code has it on conv1!!
-        self.gn3 = nn.GroupNorm(32, cmid)
-        self.conv3 = conv1x1(cmid, cout)
-        self.relu = nn.ReLU(inplace=True)
-
-        if stride != 1 or cin != cout:
-            # Projection also with pre-activation according to paper.
-            self.downsample = conv1x1(cin, cout, stride)
-
-    def forward(self, x):
-        out = self.relu(self.gn1(x))
-
-        # Residual branch
-        residual = x
-        if hasattr(self, "downsample"):
-            residual = self.downsample(out)
-
-        # Unit's branch
-        out = self.conv1(out)
-        out = self.conv2(self.relu(self.gn2(out)))
-        out = self.conv3(self.relu(self.gn3(out)))
-
-        return out + residual
-
-    def load_from(self, weights, prefix=""):
-        convname = "standardized_conv2d"
-        with torch.no_grad():
-            self.conv1.weight.copy_(tf2th(weights[f"{prefix}a/{convname}/kernel"]))
-            self.conv2.weight.copy_(tf2th(weights[f"{prefix}b/{convname}/kernel"]))
-            self.conv3.weight.copy_(tf2th(weights[f"{prefix}c/{convname}/kernel"]))
-            self.gn1.weight.copy_(tf2th(weights[f"{prefix}a/group_norm/gamma"]))
-            self.gn2.weight.copy_(tf2th(weights[f"{prefix}b/group_norm/gamma"]))
-            self.gn3.weight.copy_(tf2th(weights[f"{prefix}c/group_norm/gamma"]))
-            self.gn1.bias.copy_(tf2th(weights[f"{prefix}a/group_norm/beta"]))
-            self.gn2.bias.copy_(tf2th(weights[f"{prefix}b/group_norm/beta"]))
-            self.gn3.bias.copy_(tf2th(weights[f"{prefix}c/group_norm/beta"]))
-            if hasattr(self, "downsample"):
-                w = weights[f"{prefix}a/proj/{convname}/kernel"]
-                self.downsample.weight.copy_(tf2th(w))
+from .model_base import StdConv2d, PreActBottleneck, tf2th
 
 
 class ResNetV2Screenshot(nn.Module):
